+++ src/pages/home/index.tsx
@@ -4,6 +4,7 @@ import { HomeMobile } from "./mobile-view"
 import { useEffect, useState } from "react"
 import genresActions from "../../redux/actions/genres"
 import bookListsActions from "../../redux/actions/book-lists"
+import editionsActions from "../../redux/actions/editions"
 import type { AppDispatch } from "../../redux/store"
 
 export const Home = () => {
@@ -29,7 +30,7 @@ export const Home = () => {
             })
         );
         dispatch(bookListsActions.getAll({ limit: isDesktop ? 3 : 5 }));
-        dispatch(bookListsActions.getMostRatedBooks({ enableSuggestion: isDesktop, limit: isDesktop ? 4 : 10 }));
+        dispatch(editionsActions.getMostRatedBooks({ enableSuggestion: isDesktop, limit: isDesktop ? 4 : 10 }));
     }, [dispatch, isDesktop]);
 
     return isDesktop ? <HomeDesktop /> : <HomeMobile />;
