+++ src/redux/actions/book-lists.ts
@@ -47,75 +47,9 @@ const getByTitle = createAsyncThunk(
   },
 );
 
-const getLatestReleases = createAsyncThunk(
-  "bookLists/getLatestReleases",
-  async ({ limit }: { limit?: number }) => {
-    try {
-      const response = await axios.get(
-        `${API_URL}/book-lists/latest-releases?limit=${limit}`,
-      );
-      return response.data;
-    } catch (error: unknown) {
-      if (error instanceof Error) {
-        throw new Error(error.message);
-      }
-      throw new Error("Failed to fetch book list");
-    }
-  },
-);
-
-const getMostRatedBooks = createAsyncThunk(
-  "bookLists/getMostRatedBooks",
-  async ({
-    limit,
-    enableSuggestion,
-  }: {
-    limit?: number;
-    enableSuggestion?: boolean;
-  }) => {
-    try {
-      const response = await axios.get(
-        `${API_URL}/book-lists/most-rated?limit=${limit}&&enableSuggestion=${enableSuggestion}`,
-      );
-      return response.data;
-    } catch (error: unknown) {
-      if (error instanceof Error) {
-        throw new Error(error.message);
-      }
-      throw new Error("Failed to fetch book list");
-    }
-  },
-);
-
-const getBestRatedBooks = createAsyncThunk(
-  "bookLists/getBestRatedBooks",
-  async ({
-    limit,
-    enableSuggestion,
-  }: {
-    limit?: number;
-    enableSuggestion?: boolean;
-  }) => {
-    try {
-      const response = await axios.get(
-        `${API_URL}/book-lists/best-rated?limit=${limit}&&enableSuggestion=${enableSuggestion}`,
-      );
-      return response.data;
-    } catch (error: unknown) {
-      if (error instanceof Error) {
-        throw new Error(error.message);
-      }
-      throw new Error("Failed to fetch book list");
-    }
-  },
-);
-
 const bookListsActions = {
   getAll,
   getByTitle,
-  getLatestReleases,
-  getMostRatedBooks,
-  getBestRatedBooks,
 };
 
 export default bookListsActions;
