+++ src/redux/reducers/book-list.ts
@@ -1,13 +1,10 @@
 import { createSlice } from "@reduxjs/toolkit";
 import actions from "../actions/book-lists";
-import type { BookListI, EditionI } from "../../data-structures";
+import type { BookListI } from "../../data-structures";
 
 interface BookListsState {
   listOfBookLists: BookListI[];
   selectedBookList: BookListI | null;
-  latestReleases: EditionI[];
-  mostRatedBooks: { list: EditionI[]; suggestion: EditionI | null };
-  bestRatedBooks: { list: EditionI[]; suggestion: EditionI | null };
   status: string;
   error: string;
 }
@@ -15,9 +12,6 @@ interface BookListsState {
 const initialState: BookListsState = {
   listOfBookLists: [],
   selectedBookList: null,
-  latestReleases: [],
-  mostRatedBooks: { list: [], suggestion: null },
-  bestRatedBooks: { list: [], suggestion: null },
   status: "loading",
   error: "",
 };
@@ -66,44 +60,6 @@ const bookListsSlice = createSlice({
         state.status = "idle";
         state.error = action.error.message || "Failed to fetch book list";
         state.selectedBookList = null;
-      })
-      .addCase(actions.getLatestReleases.pending, (state) => {
-        state.status = "loading";
-      })
-      .addCase(actions.getLatestReleases.fulfilled, (state, action) => {
-        state.status = "idle";
-        state.latestReleases = action.payload;
-      })
-      .addCase(actions.getLatestReleases.rejected, (state, action) => {
-        state.status = "idle";
-        state.error = action.error.message || "Failed to fetch latest releases";
-        state.latestReleases = [];
-      })
-      .addCase(actions.getMostRatedBooks.pending, (state) => {
-        state.status = "loading";
-      })
-      .addCase(actions.getMostRatedBooks.fulfilled, (state, action) => {
-        state.status = "idle";
-        state.mostRatedBooks = action.payload;
-      })
-      .addCase(actions.getMostRatedBooks.rejected, (state, action) => {
-        state.status = "idle";
-        state.error =
-          action.error.message || "Failed to fetch most rated books";
-        state.mostRatedBooks = { list: [], suggestion: null };
-      })
-      .addCase(actions.getBestRatedBooks.pending, (state) => {
-        state.status = "loading";
-      })
-      .addCase(actions.getBestRatedBooks.fulfilled, (state, action) => {
-        state.status = "idle";
-        state.bestRatedBooks = action.payload;
-      })
-      .addCase(actions.getBestRatedBooks.rejected, (state, action) => {
-        state.status = "idle";
-        state.error =
-          action.error.message || "Failed to fetch best rated books";
-        state.bestRatedBooks = { list: [], suggestion: null };
       });
   },
 });
